Fix bad from quoting
Lines that start 'From ' are being prefixed with '>' unnecessarily. That
is needed only when writing to an mbox type file with multiple messages.
STUMP handles messages in individual files so it is not needed.

That quoting is performed by formail by default so this adds the option
to switch off the quoting.