Minimal fixes to get STUMP to work
These are minimal fixes to get STUMP to work on Linux Ubuntu 20.04.03
LTS with inn 1:1.7.2q-46build1

'reject ignore' is already handled specially but fails if there is no
'ignore' message file - fix is to not require the file.

If the RNEWS variable is set to a command with parameters - 'inews -h
-R' in my case, the command will fail if RNEWS is in quotes (two places
but the submitFailed case has not been tested).

Syntax error in processRejected causes it to fail, sending no message
and leaving the message in STUMP workspace.

My inews rejects submissions with an 'X-Trace' header so this is added
to the list of headers stripped from the input. There is a risk that
this will have continuation lines so that must be handled too.