Bug #59261: Removed 'hostname.REPLACEMENT' script
The hostname.REPLACEMENT script can never have worked as described as it
contains at least two major bugs. The use case for it (incompatibilities
with the hostname command on BSD-ish systems) also does not seem to be a
current one, as the B8MB has been running STUMP on NetBSD for many years
without encountering such problems. I have therefore deleted it, as well
as the references to it in modenv.